Move the WebSocket server out of tournament.rs
